+++ packages/in_app_purchase/lib/src/in_app_purchase/app_store_connection.dart
@@ -75,7 +75,7 @@ class AppStoreConnection implements InAppPurchaseConnection {
   @override
   Future<QueryPurchaseDetailsResponse> queryPastPurchases(
       {String applicationUserName}) async {
-    PurchaseError error;
+    IAPError error;
     List<PurchaseDetails> pastPurchases = [];
 
     try {
@@ -93,8 +93,8 @@ class AppStoreConnection implements InAppPurchaseConnection {
           ..status = SKTransactionStatusConverter()
               .toPurchaseStatus(transaction.transactionState)
           ..error = transaction.error != null
-              ? PurchaseError(
-                  source: PurchaseSource.AppStore,
+              ? IAPError(
+                  source: IAPSource.AppStore,
                   code: kPurchaseErrorCode,
                   message: transaction.error.domain,
                   details: transaction.error.userInfo,
@@ -102,14 +102,14 @@ class AppStoreConnection implements InAppPurchaseConnection {
               : null;
       }).toList();
     } on PlatformException catch (e) {
-      error = PurchaseError(
-          source: PurchaseSource.AppStore,
+      error = IAPError(
+          source: IAPSource.AppStore,
           code: e.code,
           message: e.message,
           details: e.details);
     } on SKError catch (e) {
-      error = PurchaseError(
-          source: PurchaseSource.AppStore,
+      error = IAPError(
+          source: IAPSource.AppStore,
           code: kRestoredPurchaseErrorCode,
           message: e.domain,
           details: e.userInfo);
@@ -125,7 +125,7 @@ class AppStoreConnection implements InAppPurchaseConnection {
     return PurchaseVerificationData(
         localVerificationData: receipt,
         serverVerificationData: receipt,
-        source: PurchaseSource.AppStore);
+        source: IAPSource.AppStore);
   }
 
   /// Query the product detail list.
@@ -162,8 +162,8 @@ class AppStoreConnection implements InAppPurchaseConnection {
       notFoundIDs: invalidIdentifiers,
       error: exception == null
           ? null
-          : PurchaseError(
-              source: PurchaseSource.AppStore,
+          : IAPError(
+              source: IAPSource.AppStore,
               code: exception.code,
               message: exception.message,
               details: exception.details),
@@ -216,8 +216,8 @@ class _TransactionObserver implements SKTransactionObserverWrapper {
             ..status = SKTransactionStatusConverter()
                 .toPurchaseStatus(transaction.transactionState)
             ..error = transaction.error != null
-                ? PurchaseError(
-                    source: PurchaseSource.AppStore,
+                ? IAPError(
+                    source: IAPSource.AppStore,
                     code: kPurchaseErrorCode,
                     message: transaction.error.domain,
                     details: transaction.error.userInfo,