+++ packages/in_app_purchase/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.0
+
+* [Breaking Change] Rename 'PurchaseError' to 'IAPError'.
+* [Breaking Change] Rename 'PurchaseSource' to 'IAPSource'.
+
 ## 0.1.1+3
 
 * Expanded description in `pubspec.yaml` and fixed typo in `README.md`.